requirements.txt
opencv-python==4.11.0.86
numpy==1.26.4
imagehash==4.3.2
Pillow==11.1.0
psutil==7.0.0
matplotlib==3.10.1
pytest==9.0.3